TASK: block input when not allowed to change workspace
